Migrate over to async APIs for FileIO
