+++ Tests/VaporTests/FileTests.swift
@@ -537,4 +537,20 @@ final class FileTests: XCTestCase {
             XCTAssertEqual(res.status, .badRequest)
         }
     }
+    
+    func testAsyncFileWrite() async throws {
+        let app = Application(.testing)
+        defer { app.shutdown() }
+        
+        let request = Request(application: app, on: app.eventLoopGroup.next())
+        
+        let data = "Hello"
+        let path = "/tmp/fileio_write.txt"
+        
+        try await request.fileio.writeFile(ByteBuffer(string: data), at: path)
+        defer { try? FileManager.default.removeItem(atPath: path) }
+        
+        let result = try String(contentsOfFile: path)
+        XCTAssertEqual(result, data)
+    }
 }
